fixed bug
during the for loop where it checked if i was less than the number, i was an integer so it couldn't be less than the large number hardcoded in so it overflowed and was stuck in an infinite loop.  making it a long means that the hardcoded number will have to be bigger for it too loop infinitely.